refactor: create new directory components/layouts
