+++ src/components/layouts/main-layout/index.ts
@@ -1,0 +1,1 @@
+export { MainLayout } from './MainLayout';
