+++ src/components/layouts/private-routes-layout/index.ts
